+++ src/command/metadata.rs
@@ -135,9 +135,13 @@ pub(crate) const COMMANDS: &[CommandMetadata] = &[
     metadata!("zcount", 4, READ, 1, 1, 1),
     metadata!("zincrby", 4, WRITE, 1, 1, 1),
     metadata!("zmscore", -3, READ, 1, 1, 1),
+    metadata!("zpopmax", -2, WRITE, 1, 1, 1),
+    metadata!("zpopmin", -2, WRITE, 1, 1, 1),
     metadata!("zrange", -4, READ, 1, 1, 1),
     metadata!("zrank", 3, READ, 1, 1, 1),
     metadata!("zrem", -3, WRITE, 1, 1, 1),
+    metadata!("zremrangebyrank", 4, WRITE, 1, 1, 1),
+    metadata!("zremrangebyscore", 4, WRITE, 1, 1, 1),
     metadata!("zrevrank", 3, READ, 1, 1, 1),
     metadata!("zscore", 3, READ, 1, 1, 1),
 ];